v2.9.2
Corrected a typo in the ova path so it points to the ova.jpg instead of bd720.jpg which was a copy paste error on my part. I added another switch case since they are case sensitive and that may cause certain button to not load from the repo correctly.