v3.0
I added all of A-Kuma's post images to the repo and expanded the dynamic_loading_all.js to handle more variance within spoiler images as that's what A-Kuma used for his posts on the site. Might need to fine-tune the script a bit more, but most of the new code is functional.